Use the kiwix::Reader instead of kiwix::Searcher to search suggestion.
The expected behavior from a user pov, is that suggestions return titles
starting with the searchPattern if no fulltext index is available.

`kiwix::Searcher` works only on a xapian database.
So `kiwix::Searcher::suggestions` will return results only if a embedded
fulltext is available.

This is `kiwix::Reader::searchSuggestionsSmart` that first try to do a
search with `kiwix::Searcher` and they fallback to titles starting with
the searchPattern.